requirements.txt
beautifulsoup4==4.9.1
bs4==0.0.1
cachetools==4.1.1
certifi==2020.6.20
chardet==3.0.4
dialogflow==1.0.0
gevent==20.6.2
google-api-core==1.22.1
google-auth==1.20.1
googleapis-common-protos==1.52.0
greenlet==0.4.16
grpcio==1.31.0
html2text==2020.1.16
idna==2.10
lxml==4.5.2
numpy==1.19.1
protobuf==3.13.0
pyasn1==0.4.8
pyasn1-modules==0.2.8
pytz==2020.1
requests==2.24.0
rsa==4.6
six==1.15.0
soupsieve==2.0.1
urllib3==1.25.10
websocket-client==0.57.0
zope.event==4.4
zope.interface==5.1.0
